+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log #
 All notable changes to this project will be documented in this file.
 
+## [Unreleased] ##
+- Added command-line option `-W`
+
 ## [3.4.4] - 2023-07-29 ##
 - **Added** a man page that describes how to write themes.
 - **Added** and made use of a pointer validation function.